contact: organize contact components into sections / officially finished animation for contact page
